requirements.txt
fastapi==0.115.12
GDAL==3.8.4
geopandas==1.0.1
laspy==2.5.4
matplotlib==3.10.1
numpy==1.26.4
pandas==2.1.4
Pillow==11.2.1
plotly==5.18.0
rasterio==1.4.3
scipy==1.15.2
Shapely==2.1.0
timm==1.0.15
torch==2.3.1+cu121
torchvision==0.18.1+cu121
